requirements.txt
appdirs==1.4.4
astor==0.8.1
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
dash==3.1.1
dash-bootstrap-components==2.0.3
decorator==5.2.1
entsoe-py==0.7.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.21.1
Flask==3.1.1
fonttools==4.58.5
forex-python==1.9.2
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_core==5.8.1
kiwisolver==1.4.7
lxml==6.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
narwhals==1.44.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.0.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pandas-datareader==0.10.0
pandaSDMX==1.10.0
pandasgui==0.2.15
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pydantic==1.10.22
Pygments==2.19.2
pynput==1.8.1
pyobjc-core==11.1
pyobjc-framework-ApplicationServices==11.1
pyobjc-framework-Cocoa==11.1
pyobjc-framework-CoreText==11.1
pyobjc-framework-Quartz==11.1
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.17
python-dateutil==2.9.0.post0
pytz==2025.2
qtstylish==0.1.5
referencing==0.36.2
requests==2.32.4
retrying==1.4.0
rpds-py==0.26.0
scipy==1.13.1
simplejson==3.20.1
six==1.17.0
soupsieve==2.7
stack-data==0.6.3
statsmodels==0.14.4
traitlets==5.14.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
wordcloud==1.9.4
zipp==3.23.0
